requirements.txt
streamlit==1.22.0
huggingface_hub==0.15.1
google-generativeai==0.1.0
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.2
matplotlib==3.7.1
nltk==3.8.1
PyPDF2==3.0.1
faiss-cpu==1.7.3
langchain==0.0.1